+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed `game` command not working when the specified channel contained the @ symbol at the beginning.
 - Fixed strimertul crashing at start if the database folder didn't exist and a database driver was not manually specified
+- Fixed strimertul crashing after a minute of being open if Twitch was not configured.
 
 ## [2.1.1] - 2022-03-24